LayoutTests/fast/dom/dom-serialize-namespace-expected.txt

This test verifies that the namespace on a node gets serialized if present even if the node is not in the DOM tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS s.serializeToString(node).replace(/"/g,"'") is "<test xmlns='http://example.com/'/>"
PASS s.serializeToString(doc).replace(/"/g,"'") is "<test xmlns='http://example.com/'/>"
PASS successfullyParsed is true

TEST COMPLETE
